注入防线三层接入 + 测试脚本归入 tests/ 目录
注入检测（三层防线）：
- model/injection_detector.py：ONNX 推理 + 三级分级(0.05/0.2/0.9)
  + 恶意值 Redis 原子存储（加分+衰减+禁言合并单 Lua 脚本 EVAL）
- routes/ai_chat.py：图外检测——禁言/硬拦直接拒绝，suspect/warn 经 UserContext 传图
- chat_node：当轮旁白（方案K，AIMessage 不落 checkpoint，明确指向本条消息）
- out_node：输出侧系统提示词指纹脱敏（PROMPT_FINGERPRINTS）

测试整理：
- 测试脚本/测试数据归入 tests/（git mv 保留历史）
- load_config 根路径改为模块锚点（python -c 可用，测试脚本可放任意目录运行）

注入防御训练管线：
- 数据集构建/质量检查（build_injection_dataset.py / check_injection_dataset.py）
- Kaggle 训练 notebook + 生成器（train_injection_classifier.ipynb）
- ONNX 本地验证脚本（verify_onnx.py，tokenizers 0.22 与训练环境对齐）

其他：requirements 增 onnxruntime/tokenizers；.gitignore 增 model 大文件；README 目录树更新

+++ SPO/data_model.py
@@ -63,8 +63,8 @@ class CouponInfo(Base):
 
 class UserOrderItem(Base):
     __tablename__ = 'ordercart'
-    id = Column(Integer, primary_key=True, comment='主键id')
-    orderId = Column(Integer, default=0, comment='订单ID')
+    id = Column(BigInteger, primary_key=True, comment='主键id')
+    orderId = Column(BigInteger, default=0, comment='订单ID')
     userId = Column(BigInteger, default=0, comment='用户ID')
     goodsId = Column(Integer, default=0, comment='商品ID')
     quantity = Column(Integer, default=0, comment='购买的数量')
@@ -86,7 +86,7 @@ class UserOrderJson(BaseModel):
 
 class OrderInfo(Base):
     __tablename__ = 'order'
-    orderId = Column(Integer, primary_key=True, comment='订单ID')
+    orderId = Column(BigInteger, primary_key=True, comment='订单ID')
     userId = Column(BigInteger, default=0, comment='用户ID')
     totalAmount = Column(Float, default=0, comment='订单总金额(单位：元)')
     receiptAddress = Column(String(255), default='', comment='订单收货地址')
